changed models to typescript
