+++ packages/backend/server/models/counter.ts
@@ -1,22 +1,47 @@
-
-export {}
+export {};
 
 const name = 'Counter';
 const tableName = 'counters';
-
 const selectableProps = ['id', 'counter'];
-
-const createGuts = require('../helpers/model-guts');
+const timeout = 1000;
 
 module.exports = (knex) => {
-  const guts = createGuts({
-    knex,
+  const create = (props) => {
+    delete props.id; // not allowed to set `id`
+
+    return knex
+      .insert(props)
+      .returning(selectableProps)
+      .into(tableName)
+      .timeout(timeout);
+  };
+
+  const findAll = () =>
+    knex.select(selectableProps).from(tableName).timeout(timeout);
+
+  const findById = (id) =>
+    knex.select(selectableProps).from(tableName).where({ id }).timeout(timeout);
+
+  const update = (id, props) => {
+    delete props.id; // not allowed to set `id`
+
+    return knex
+      .update(props)
+      .from(tableName)
+      .where({ id })
+      .returning(selectableProps)
+      .timeout(timeout);
+  };
+
+  return {
     name,
     tableName,
     selectableProps,
-  });
-
-  return {
-    ...guts,
+    timeout,
+    create,
+    findAll,
+    findById,
+    update,
   };
 };
+
